refactor: :truck: separate a file and move to sub package
