+++ app/src/main/java/com/example/k42un0k0smoke/implement/repository/QuitResultRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.example.k42un0k0smoke.implement.repository
 import androidx.annotation.WorkerThread
 import com.example.k42un0k0smoke.infrastructure.room.QuitResultDao
 import com.example.k42un0k0smoke.infrastructure.room.QuitResultDto
-import com.example.k42un0k0smoke.model.QuitResult
+import com.example.k42un0k0smoke.model.quit_result.QuitResult
 import com.example.k42un0k0smoke.repository.QuitResultRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
