+++ app/src/main/java/com/example/k42un0k0smoke/model/quit_result/QuitResultTotalSavings.kt
@@ -1,14 +1,7 @@
-package com.example.k42un0k0smoke.model
+package com.example.k42un0k0smoke.model.quit_result
 
 import com.example.k42un0k0smoke.lib.DurationUtil
 import java.time.LocalDateTime
-
-data class QuitResult(
-    val uid: Int?,
-    val startAt: LocalDateTime,
-    val endAt: LocalDateTime,
-    val totalSavings: QuitResultTotalSavings
-)
 
 data class QuitResultTotalSavings(val value: Long) {
     override fun toString(): String {
