+++ app/src/main/java/com/example/k42un0k0smoke/modules/quit_results/QuitResultsViewModel.kt
@@ -3,7 +3,7 @@ package com.example.k42un0k0smoke.modules.quit_results
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.example.k42un0k0smoke.model.QuitResult
+import com.example.k42un0k0smoke.model.quit_result.QuitResult
 import com.example.k42un0k0smoke.repository.QuitResultRepository
 import javax.inject.Inject
